pyproject.toml
[project]
name = "behavioral-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "anthropic>=0.69.0",
    "google-genai>=1.41.0",
    "openai>=2.2.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pyyaml>=6.0.3",
    "streamlit[charts]>=1.50.0",
]
